feat(oracle): move code to racle layer
