+++ src/dialects/oracle/data-types.js
@@ -4,6 +4,7 @@
 
 const moment = require('moment');
 const momentTz = require('moment-timezone');
+const Utils = require('../../utils');
 
 module.exports = BaseTypes => {
   const warn = BaseTypes.ABSTRACT.warn.bind(
@@ -463,6 +464,8 @@ module.exports = BaseTypes => {
   class VECTOR extends BaseTypes.VECTOR {
     constructor(dimension, format) {
       super(dimension, format);
+      this._length = typeof dimension === 'object' && dimension.dimension || dimension;
+      this._format = typeof dimension === 'object' && dimension.format || format;
     }
     toSql() {
       if (this._length && this._format) {
@@ -478,6 +481,33 @@ module.exports = BaseTypes => {
       return { type: oracledb.DB_TYPE_VECTOR };
     }
 
+
+  }
+
+  BaseTypes.VECTOR.prototype.cosineDistance = function cosineDistance(column, value, sequelize) {
+    return distance('COSINE_DISTANCE', column, value, sequelize);
+  };
+  
+  BaseTypes.VECTOR.prototype.innerProduct = function innerProduct(column, value, sequelize) {
+    return distance('INNER_PRODUCT', column, value, sequelize);
+  };
+
+  BaseTypes.VECTOR.prototype.l1Distance = function l1Distance(column, value, sequelize) {
+    return distance('L1_DISTANCE', column, value, sequelize);
+  };
+  
+  BaseTypes.VECTOR.prototype.l2Distance = function l2Distance(column, value, sequelize) {
+    return distance('L2_DISTANCE', column, value, sequelize);
+  };
+  
+  BaseTypes.VECTOR.prototype.vectorDistance = function vectorDistance(column, value, sequelize) {
+    return distance('vector_distance', column, value, sequelize);
+  };
+
+  function distance(distanceType, column, value, sequelize) {
+    const quotedColumn = column instanceof Utils.Literal ? column.val : sequelize.dialect.queryGenerator.quoteIdentifier(column);
+    const val = `VECTOR('[${value}]', ${value.length})`;
+    return `${distanceType}(${quotedColumn}, ${val})`;
   }
 
   return {
